--- MovieQuiz/Presentation/MovieQuizViewControllerMock.swift
@@ -1,36 +1,0 @@
-//
-//  MovieQuizViewControllerMock.swift
-//  MovieQuiz
-//
-//  Created by Oschepkov Aleksandr on 13.01.2026.
-//
-
-import UIKit
-
-final class MovieQuizViewControllerMock: MovieQuizViewControllerProtocol{
-    func showLoadingIndicator() {
-
-    }
-    
-    func hideLoadingIndicator() {
-    
-    }
-    
-    func configureImageView(isCorrect: Bool) {
-    
-    }
-    
-    func show(quiz step: QuizStepViewModel) {
-        
-    }
-    
-    func showNextQuestionOrResults() {
-        
-    }
-    
-    func showNetworkError(message: String) {
-        
-    }
-    
-    
-}
